Fix TCP TransferBytes

+++ BarbaTunnel/BarbaTunnel/BarbaBuffer.h
@@ -23,8 +23,8 @@ public:
 	void append(T value) {buf.push_back(value);}
 	bool empty() {return buf.empty();}
 	void clear() {return buf.clear();}
-	T operator[](int nIndex) const { return buf[nIndex]; }
-	T& operator[](int nIndex) { return buf[nIndex]; }
+	T operator[](size_t nIndex) const { return buf[nIndex]; }
+	T& operator[](size_t nIndex) { return buf[nIndex]; }
 
 protected:
 	std::vector<T> buf;
